+++ build.sh
@@ -1,14 +1,15 @@
 #!/usr/bin/env bash
 
 # Get composer
-EXPECTED_SIGNATURE="$(wget -q -O - https://composer.github.io/installer.sig)"
-php -r "copy('https://getcomposer.org/installer', 'composer-setup.php');"
-ACTUAL_SIGNATURE="$(php -r "echo hash_file('sha384', 'composer-setup.php');")"
+# Locked to version 1.X.X to preserve PHP 5.6 compatibility
+EXPECTED_SIGNATURE="e494bb438e44b9e4782c16940b229a8c46ea8a3baa9b908bf9db310cd0171ee2"
+php -r "copy('https://getcomposer.org/download/1.10.16/composer.phar', 'composer.phar');"
+ACTUAL_SIGNATURE="$(php -r "echo hash_file('sha256', 'composer.phar');")"
 
 if [ "$EXPECTED_SIGNATURE" != "$ACTUAL_SIGNATURE" ]
 then
-    >&2 echo 'ERROR: Invalid installer signature'
-    rm composer-setup.php
+    >&2 echo 'ERROR: Invalid composer version'
+    rm composer.phar
     exit 1
 fi
 
@@ -18,17 +19,42 @@ rm composer-setup.php
 
 # Remove vendor directory
 rm -rf vendor
+rm -rf build
 
 # Run composer install
 php composer.phar install
 
+# Require and run php-scoper
+# Locked to version compatible with version of composer
+php composer.phar global require humbug/php-scoper "0.12.3"
+COMPOSER_BIN_DIR="$(composer global config bin-dir --absolute)"
+"$COMPOSER_BIN_DIR"/php-scoper add-prefix
+
+# Dump composer autoload for build folder
+php composer.phar dump-autoload --working-dir build --classmap-authoritative
+
 # Remove composer
 rm composer.phar
 
-# Zip contents of folder to onpay folder in a zip file
+# Remove existing build zip file
 rm onpay.zip
+
+# Rsync contents of folder to new directory that we will use for the build
 rsync -Rr ./* ./onpay
+
+# Remove directories and files from newly created directory, that we won't need in final build
+rm -rf ./onpay/vendor
+rm ./onpay/build.sh
+rm ./onpay/composer.json
+rm ./onpay/composer.lock
+
+# Replace require file with build version
+rm ./onpay/require.php
+mv ./onpay/require_build.php ./onpay/require.php
+
+# Zip contents of newly created directory
 zip -r onpay.zip ./onpay
 
 # Clean up
 rm -rf onpay
+rm -rf build